Simplify the script
since there is no need for sha256sum, we can remove it. Then this mean
there is no need for IMAGE_NAME, so we can remove it. And so we can
remove FILESYSTEM. And since FILES should be files by default (as that's
default on
https://openwrt.org/docs/guide-user/additional-software/imagebuilder ),
we can change it as well.